Introduce selection system, gradient decorations, and refactor component styling logic
- Release version 1.0.6 and bump minimum Flutter dependency to 3.38.0.
- Introduce the `M3ESelection` system, including `M3ESelectionController`, `M3ESelectionAppBar`, and automatic highlight support for card and dismissible lists.
- Add support for gradient decorations (`background`, `foreground`, `overlay`, and `outline`) across `M3EButton`, `M3EFab`, `M3EIconButton`, and `M3ESegmentedButton`.
- Update `M3EBadge` with `alignment` support and improved sizing logic that respects the child's box.
- Refactor `M3EFab` and `M3EExtendedFab` to modularize surface building and content decoration logic.
- Enhance `M3ETextField` with `inputFormatters` support and fix layout stability issues during focus state changes.
- Modernize `M3ESplitButton` with `m3eMenuBuilder` for rich menu trees and support for segment-specific gradient overrides.
- Refactor `M3ESplitButtonDecoration` and `M3EToggleButtonDecoration` for better maintainability, including splitting toggle decorations into a separate part file.
- Update the example gallery with a new playground structure and a comprehensive theme configuration screen for dynamic color and palette testing.
