+++ example/lib/pages/section_host_page.dart
@@ -122,9 +122,7 @@ class _WideDetail extends StatelessWidget {
   Widget build(BuildContext context) {
     return Column(
       crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: <Widget>[
-        Expanded(child: child),
-      ],
+      children: <Widget>[Expanded(child: child)],
     );
   }
 }
