+++ example/lib/pages/theme_config_page.dart
@@ -19,27 +19,32 @@ class ThemeConfigPage extends StatelessWidget {
     return Scaffold(
       body: ColoredBox(
         color: theme.colorScheme.surface,
-        child: Column(
-          children: <Widget>[
-            M3EAppBar.top(
-              titleText: 'Theme',
-              leading: M3EIconButton(
-                icon: const Icon(M3EIcons.arrow_back),
-                tooltip: 'Back',
-                onPressed: () => Navigator.of(context).maybePop(),
+        child: Padding(
+          padding: EdgeInsets.only(
+            bottom: MediaQuery.viewPaddingOf(context).bottom,
+          ),
+          child: Column(
+            children: <Widget>[
+              M3EAppBar.top(
+                titleText: 'Theme',
+                leading: M3EIconButton(
+                  icon: const Icon(M3EIcons.arrow_back),
+                  tooltip: 'Back',
+                  onPressed: () => Navigator.of(context).maybePop(),
+                ),
               ),
-            ),
-            Expanded(
-              child: ListView(
-                padding: const EdgeInsets.fromLTRB(16, 8, 16, 24),
-                children: <Widget>[
-                  _toggles(theme, settings),
-                  const SizedBox(height: 24),
-                  _seeds(theme, settings),
-                ],
+              Expanded(
+                child: ListView(
+                  padding: const EdgeInsets.fromLTRB(16, 16, 16, 24),
+                  children: <Widget>[
+                    _toggles(theme, settings),
+                    const SizedBox(height: 24),
+                    _seeds(theme, settings),
+                  ],
+                ),
               ),
-            ),
-          ],
+            ],
+          ),
         ),
       ),
     );
